organização

+++ test.py
@@ -5,7 +5,7 @@ from sklearn.model_selection import cross_val_score
 import pickle
 
 Xtest = np.loadtxt("Xtest.csv", delimiter=",")
-# Xtest = Xtest.astype(int)
+Xtest = Xtest.astype(int)
 print(Xtest)
 print(Xtest.shape)
 ytest = np.loadtxt("ytest.csv", delimiter=",")
@@ -16,7 +16,7 @@ print(ytest.shape)
 model = pickle.load(open("model.mlp", 'rb'))
 
 Xtrain = np.loadtxt("Xtraining.csv", delimiter=",")
-# Xtrain = Xtrain.astype(int)
+Xtrain = Xtrain.astype(int)
 scaler = StandardScaler()
 scaler.fit(Xtrain)
 ypredicted = model.predict(scaler.transform(Xtest))
